[DF-7571] BlockAllocator sources added to framework projects
Former-commit-id: c35fb39fb18c94e01343cec0e8f4a4bfe440bc7e

+++ Sources/Internal/Base/BlockAllocator.h
@@ -82,7 +82,7 @@ BlockAllocator<blockSize, numBlocks>::BlockAllocator()
 }
 
 template <uint32 blockSize, uint32 numBlocks>
-char* BlockAllocator<blockSize, numBlocks>::Alloc(size_t sz)
+char* BlockAllocator<blockSize, numBlocks>::Alloc(uint32 sz)
 {
     DVASSERT(sz <= blockSize);
